float64

+++ xcube_resampling/rectify.py
@@ -349,13 +349,13 @@ def _get_pixel_target_ij(
         scale: np.ndarray,
     ) -> np.ndarray:
         """Compute fractional source indices for a single target tile."""
-        target_x = tgt_x_block[0, :].astype(np.float32)
-        target_y = tgt_y_block[:, 0].astype(np.float32)
+        target_x = tgt_x_block[0, :].astype(np.float64)
+        target_y = tgt_y_block[:, 0].astype(np.float64)
 
         nx_tgt = target_x.size
         ny_tgt = target_y.size
 
-        idx_frac = np.full((2, ny_tgt, nx_tgt), np.nan, dtype=np.float32)
+        idx_frac = np.full((2, ny_tgt, nx_tgt), np.nan, dtype=np.float64)
 
         ny_src, nx_src = src_x_block.shape
 
@@ -365,7 +365,7 @@ def _get_pixel_target_ij(
         quad_ij0 = np.stack((ix0, iy0)).astype(np.int32)
 
         # Quad corner coordinates
-        src_coords = np.stack((src_x_block, src_y_block)).astype(np.float32)
+        src_coords = np.stack((src_x_block, src_y_block)).astype(np.float64)
         quad_corners = np.stack(
             (
                 src_coords[:, quad_ij0[1], quad_ij0[0]],
@@ -384,7 +384,7 @@ def _get_pixel_target_ij(
             return idx_frac
 
         # Bounding target pixel indices per quad
-        offset = np.array([target_x[0], target_y[0]], dtype=np.float32)
+        offset = np.array([target_x[0], target_y[0]], dtype=np.float64)
         tgt_bbox = np.floor(
             (quad_corners - offset[:, None, None]) / scale[:, None, None]
         ).astype(np.int32)
@@ -422,7 +422,7 @@ def _get_pixel_target_ij(
     # Build block-aligned target coordinate views
     tgt_x = target_gm.x_coords.data
     tgt_y = target_gm.y_coords.data
-    scale = np.array([tgt_x[1] - tgt_x[0], tgt_y[1] - tgt_y[0]], dtype=np.float32)
+    scale = np.array([tgt_x[1] - tgt_x[0], tgt_y[1] - tgt_y[0]], dtype=np.float64)
     tgt_x = da.stack([tgt_x] * len(tgt_y.chunks[0]), axis=0)
     tgt_y = da.stack([tgt_y] * len(tgt_x.chunks[1]), axis=1)
 
@@ -433,7 +433,7 @@ def _get_pixel_target_ij(
         src_x_coords,
         src_y_coords,
         scale,
-        dtype=np.float32,
+        dtype=np.float64,
         chunks=(2, tgt_y.chunks[0], tgt_x.chunks[1]),
     )
 
@@ -539,7 +539,7 @@ def _xy_bbox_block(x_coords: np.ndarray, y_coords: np.ndarray):
             x_edges.max(),
             y_edges.max(),
         ],
-        dtype=np.float32,
+        dtype=np.float64,
     )
     return bbox[:, None, None]
 
@@ -549,7 +549,7 @@ def _get_xy_bboxes(gm_2d: GridMapping):
         _xy_bbox_block,
         gm_2d.x_coords.data,
         gm_2d.y_coords.data,
-        dtype=np.float32,
+        dtype=np.float64,
         chunks=(4, 1, 1),
     )
